client/requirements.txt
pyserial~=3.5
pytest~=8.3.3